Display bumpkin power in the HUD
