+++ src/features/portal/bumpkin-fight-club/components/Hud.tsx
@@ -7,19 +7,14 @@ import { Balance } from "components/Balance";
 import { BlockBucks } from "features/island/hud/components/BlockBucks";
 import { BumpkinAvatar } from "features/island/hud/components/BumpkinProfile";
 import { PIXEL_SCALE } from "features/game/lib/constants";
-import { SUNNYSIDE } from "assets/sunnyside";
-import worldIcon from "assets/icons/world.png";
 
 import { PortalContext } from "../lib/Provider";
-import { goHome } from "../lib/utils";
+import { HomeButton } from "./HomeButton";
+import { Power } from "./Power";
 
 export const Hud: React.FC = () => {
   const { portalService } = useContext(PortalContext);
   const [portalState] = useActor(portalService);
-
-  const travelHome = () => {
-    goHome();
-  };
 
   return (
     <>
@@ -29,6 +24,7 @@ export const Hud: React.FC = () => {
           aria-label="Hud"
           className="absolute z-40"
         >
+          <Power power={portalState.context.bumpkinPower} />
           <BumpkinAvatar bumpkin={portalState.context.state.bumpkin} />
           <Balance balance={portalState.context.state.balance} />
           <BlockBucks
@@ -45,36 +41,7 @@ export const Hud: React.FC = () => {
               width: `${PIXEL_SCALE * 22}px`,
             }}
           >
-            <div
-              id="deliveries"
-              className="flex relative z-50 justify-center cursor-pointer hover:img-highlight"
-              style={{
-                width: `${PIXEL_SCALE * 22}px`,
-                height: `${PIXEL_SCALE * 23}px`,
-              }}
-              onClick={(e) => {
-                e.stopPropagation();
-                e.preventDefault();
-                travelHome();
-              }}
-            >
-              <img
-                src={SUNNYSIDE.ui.round_button}
-                className="absolute"
-                style={{
-                  width: `${PIXEL_SCALE * 22}px`,
-                }}
-              />
-              <img
-                src={worldIcon}
-                style={{
-                  width: `${PIXEL_SCALE * 12}px`,
-                  left: `${PIXEL_SCALE * 5}px`,
-                  top: `${PIXEL_SCALE * 4}px`,
-                }}
-                className="absolute"
-              />
-            </div>
+            <HomeButton />
           </div>
         </div>,
         document.body
